Migrate stock_inventory_hierarchical to v8
Migration/adaptation to Odoo v8
===============================
Use `stock.inventory.location_id` (now it's standard)
Move the "generate inventories" wizard to stock_inventory_hierarchical (drop the field exhaustive)
Use v8 API
Drop the exception `HierarchicalInventoryException` in favor of the new standard exception `ValidationError`
Place the sub-inventories outside the notebook on the main inventory form (the notebook is hidden on draft inventories)
Update the translation template and translation into French.

Improvements
============
When creating a hierarchy of inventories using the wizard with the "views only" feature, make sure it matches the view hierarchy. In v7 it would ignore views which were inside internal locations, but this is more common in v8 because a location can't be the source or destination of any stock moves.
Automatically open the sub-inventories when the main inventory is opened. This avoids having to compute useless lines for the main inventory and then re-computing the same lines when opening the sub-inventories
Automatically cancel (set to draft) the sub-inventories when the main inventory is canceled.
Add warnings when clicking the buttons "Start inventory" and "Cancel inventory". The two changes above are a bit invasive, so we must make sure users know what's going to happen.
Speed improvement: defer the computation of `parent_left`/`parent_right` to the end of the wizard.
Add documentation (cf README.rst)

Coding standards
================
Clearly define (and strictly enforce) the consitency of the states of the inventories with their sub-inventories.
Rewrite the tests in API v8, improve the coverage. Drop the demo data in favor of a much larger test dataset, automatically generated during the tests.
Adopt the newest OCA standards for licences, README
Bump version to 2.0
Update the screen captures