[IMP] stock_available_mrp: Easier and faster BoM finding
Before this patch, the method was looping through found BoMs in each product, which produced an exponential double loop.

In a first attempt, I tried just avoiding the double loop, but then I noticed that with the inclusion of https://github.com/odoo/odoo/pull/33967, just calling upstream's method was easier, safer (takes more conditions into account, such as the company) and faster (due to cache, which also benefited further calls).

So here is the patch.