Support split of moves for sourcing from different zones
When a move has several move lines because it is sourced from different
locations, and these locations come from different zones, or from a zone
and another location which is not a zone, we have to split the move in
as many moves as we have zones, and chain them properly.
This is needed because otherwise, the move for, for instance taking
goods from a shelf would have to wait on the move taking from the
Higbay which has been added in front of it. We expect that we can
already process the Shelf one.

The algorithm is to find all the zones of a move, split the move if any.
But to do so, we have to unreserve the move first. As we want to keep
the same quantities from the same locations (eg. 6 from the highbay bin
1), when we reserve again the quants, we have to force the reservation
system to take the goods from the same location than we had originally
(otherwise, the quantities that we used to split the move may change and
we are back to the beginning).