Replace SQL unique constraint by a python one
The new constraints prevents having 2 locations with the same name in
the same zone. The SQL UNIQUE constraint was blocking the installation
of some addons due to demo data creating, by default, some locations
with the same name for the warehouses' locations.

Alternative: we could create an INDEX UNIQUE (name, location_id) WHERE
picking_zone_id IS NOT NULL
But it depends if we want to check the whole tree or not.