Change 'empty_bins' algorithm to sort qtys desc
The goal of the 'Empty Bins' removal strategy is to take goods from
a location only if it makes it empty afterwards. Ordering the quantities
from smallest to largest means we would empty more bins, but now we
prefer instead to reduce the number of operations. Ordering them
from largest to smallest will minimize the number of ops.