Fix flaky test in cubiscan measures
The module product_packaging_type creates a packaging type "Default
Type" with sequence 1. In the test, the "unit" is created with sequence
1 too, which makes the ordering of the packagings in the test
"test_cubiscan_measures" undeterministic.

Note: "unit" should not be a packaging type, although the wizard should
allow to change the width, length and weight of the product too.