Change sync of destination to happen from next move
First big change is that now, the 'sync' flag has to be set on the
"destination picking type", not the current one.

When we want to "regroup" moves from different transfer to the same
place for a packing, we want to activate this feature on the said
packing operation type.

Also, instead of changing the destination of move lines every time we change
a move line, we change it the first time a move is set to done to go
towards this location. We don't need to do it until the first line is
effectively moved.