[FIX] sale_stock_info_popup: Modify list editable support
Previous this commit, if the user click on a cell (after the cell created by
this widget) Odoo gets a wrong index value and try to active one of the following
cells.
With this commit the cell created by this widget is treated like a button..
that are fully compatible with tree views.