[ADD] stock_zipcube
Zipcube is a device produced by Bosche which allows weighing and measuring packages
in a way similar to Cubiscan devices.

This PR adds a stock_zipcube module which builds on the stock_cubiscan module to
add support for Zipcube devices.

Future plan:
* extract common bits in a base_stock_measurement_scale module to allow installing stock_zipcube without pulling the cubiscan dependencies

[IMP] stock_zipcube: show pop-up after scanning to press refresh

Before, after scanning with the zipcube, the measurements were
sent to Odoo but the wizard was not refreshed, so this was
confusing to the user. With this PR, a pop-up is shown to the
user that opened the wizard, warning he/she to press the button
that refreshes the screen.

Apart from that, some bugs have been corrected:
* Cubiscan-only fields were required by zipcube (device_address,
  port and timeout). So now they are required only if the device
  is cubiscan (and not zipcube). This has required moving the
  requirement attribute from the field (model) to the view.
* Pressing the button that sets the line that is going to receive
  the measures didn't set correctly the scanning device for the
  package of the line.
* The check for the ports is done only if the device is Cubiscan.
* Measure key max_weight is not supported, so has been changed in
  the code that reads it to weight.