[ADD] stock_measuring_device
This is a rework of the stock_cubiscan module to allow supporting
different kind of measuring devices, with different communication
protocols. We use components from OCA/connector to abstract the
implementation of the backend (some devices are synchronous, others are
asynchronous...).

Others PR to come for stock_measuring_device_cubiscan and
stock_measuring_device_zippcube.