[FIX] stock_move_location: Improve tests to reflect real usage
When setting the product quantities, force a small difference in the creation time.
This is important because Odoo then sorts the quantities by in_date, and when they all have the same time, sorting is a bit left to chance, and random integration errors are introduces
This is also closer to a real flow, as we woulnd't create 2 quantities at the same time.