[IMP] stock_move_location: add option to propose available quantities
This commit adds the available quantity column in the move location wizard line,
and allows users to configure in the operation type as necessary to make the available quantity
(non-reserved quantity) the default move quantity instead of the max quantity (quantity on hand).

With this, the risk of unexpectedly moving some reserved quants can be mitigated.