[FIX] stock_mts_mto_rule: don't update route
Odoo declares routes as noupdate records so the user has the control
over them. Another side effect of updatable route records is on
initializing the record when OpenUpgrade is used to upgrade, which in a
multicompany environment with mixed companies in the products could lead
to a false company restriction error (maybe due to a bad declaration of
restrictions in the product_ids field).

TT31601