[FIX] scrap_reason_code (Location Not Set)
Currently the reason code and scrap location on the code has to be set otherwise an error is generated stating the location is required.
This fixes that so if the reason code is not used or there is not a location on the code then the location on the default scrap location on the order is used, otherwise it will use the code's location.

Update Test