[FIX] stock_reserve_rule: allow removal strategy based on lots and/or owner
The lot is ignored in _apply_strategy which makes it not possible to
create removal rules based on it. A practical example is to create a
rule with a quant_domain on lot_id.use_date (Best Before Date). Even
though the quants will be filtered, the lots wont be taken into
consideration in _apply_strategy and so random lots will be reserved.
With tis the signature for the _apply_strategy_%s methods is changed
to require returning the lot also or set it to None if it is not used.

The same rationale applies for the owner of the quant.