[IMP] stock_orderpoint_generator: add  new fields
once "Automatic Generation" is enabled, let's display a flag "use product domain": if enabled, a domain for products is displayed, and table "products" is hidden.
With this feature, user does not need to select products one by one, but can apply a domain of products for which a rule should be created automatically.