[IMP] stock_inventory: new small features and ux improvements
Allow to create stock.quants making the quants smart button not invisible when quantity is equal to 0. (Removing one of the invisible conditions).

[IMP] stock_inventory: adds exclude_sublocation flag

[IMP] stock_inventory: Add flag to autocomplete adjustment when fully done.

[IMP] stock_inventory

Add company to adjustments to avoid multi-company errors. Adds information fields to tree view. Also adds 'Assign to' in stock inventory that propagates to quants. Also propagates the date field.