[IMP] stock_move_location: Improved usage of wizard
Ensure that the picking types available for use are for internal transfers.

Not change the picking type if selected before selecting an origin location of the same location default.

Add the exclusion of reserved quantities as optional, to avoid trying to transfer more than really available but also maintaining the option to transfer all even the reserved.

Add new field to show all the quantity of the line to see the difference of the available and reserved.