[IMP] stock_move_location: Improve UI and add hooks
The sql query is replaced by the read_group method from the orm. This
allows the use of odoo domain and the definition of a new method that can
be overriden to get the domain to retrieve the quants.  Extract the
initialization logic for the stock_move_location_line_ids field into a
dedicated method to allow to be called wy others addons.

Improve the wizard form:

 * Toggle fields are now properly displayed
 * Add a placeholder group to be used by specialized addon to add additional filters