[IMP] stock_location_product_restriction: Allow to put products in location if being emptied
In day-to-day stock operations, operators will take products from a location and
often empty that location.

But, if the movement quantity is already changed ('qty_done'), the picking is not
yet validated.

In parallel, some operators want to fill in the location with another product as
it is already physically void.